Allow to retreive the local path, note:
this is only working for local storage, if there are future
storages this might not work. There could be code to copy
the file locally temporary if the local storage location
is required. But maybe the carrier module should be rewritten
so that there is always a local storage plus optional
external storage locations (uploaded to a CDN).
For fingerprinting it is necessary to have a local copy.